refactor: split sidebar into pure helpers and GAS-coupled entry point
Move showAIPanel, hideAIPanel, dispatchTool, runAI, and init() from
sidebar.ts into a new sidebar-entry.ts, mirroring the server-side
pattern where index.ts (GAS-coupled) is excluded from coverage and
pure modules (api.ts, utils.ts, etc.) hold high thresholds.

sidebar-entry.ts is excluded from collectCoverageFrom; sidebar.ts now
hits 100% statements/functions and 86% branches against its existing
95/81/95 thresholds with no istanbul-ignore annotations needed.
