+++ src/Sidebar.html
@@ -70,7 +70,8 @@
             <div class="field-group">
                 <span class="field-label">Output column <span class="required">*</span></span>
                 <div id="output-col" class="tag-list"></div>
-                <input id="new-col-input" type="text" class="text-input" placeholder="ai_column_name" value="ai_" style="display:none">
+                <input id="new-col-input" type="text" class="text-input" placeholder="ai_column_name" value="ai_"
+                    style="display:none">
             </div>
 
             <div class="field-group">
